Fixed first cell not usable when selected
I noticed that you cannot select the very first cell (index 0) and
combine it with another cell to cross them out. I realized the problem
was that `null` and `0` are both falsy. I updated the check in the App
component to specifically check for `null`. Now you can select the first
cell and combine it with another one.