Moved babel config; added options
In this comit I moved the `.babelrc` to `babel.config.js`. This way I
can use the syntax highlighting of my editor and auto formatting on that
file. Also I added the target option to build for the current node
version. This way using ES6 generators in jest test won't produce an
error.