Row based rendering
In this commit I changed the way of rendering. Instead of a giant grid
layout containing all cells directly, we now have rows, that contain
their individual cells.

This serves as preparation for bringing animation to the game.

I extended the Games interface by some row functionality. Doing so I
completely renamed everything in there. I also restructured the tests as
well. Almost got lost in there.

I added the components Field and Row. Field for encapsulating the loginc
used for rendering the game field. This is where the animation will be
handled. The Row conponent cares about rendering it cells. Possibly
animating the cells as well.

I adapted the css to now use a combination of flex and grid layout. flex
for the rows, and grid for the cells in the rows.