Added configuration for test summary
In this commit I added configuration for having a test summary displayed
in circle ci.

I installed the `jest-junit` package and implemented it as reporter for
jest in the `jest.config.js` and in the circle ci config. To not have
the test reports pollute my repo I excluded them in the gitignore.