Firebase hosting configuration
With this commit I used the CLI to initialize the project as a firebase
project. I generated these config files which should be everything I
need for deployment on firebase.

This is the preparation for using the firestore to store unfinished
games.

Changes I made:
Updated the `.gitignore` to exclude firebase hosting cache files.
Updated the `webpack.config.js` to accomodate for the now missing
`/conu/` subdirectory when hosted on firebase.
Improved build script so that it removes all old artifacts before
building the new ones.
And finally I ran a build and bundled everthing again.