Renamed `Game` to `GameState`
In this commit I renamed the `Game.js` file and it's test to
`GameState.js`. This is to more accurately reflect its task in the
project and also in order to prepare for extracting the game view into
its own compoent. This means the code currently in the `App.jsx` file
that manages the displayment of the games state.