Designed redirect notice component
In this commit I took the time to design a small component indicating
the new location of the game. This includes a small message as well as
a button asking the user to go to the now official url.

This is further preparation for an url based redirection, so that the
game is only playable on conu.app. It is not really important at this
stage to have this redirect - at least not for SEO purposes. I jsut want
to have a clear user experience and promote the official url.