Webpack DefinePlugin for setting env vars
With this commit the webpack configuration was adapted to pass in
environment variables into the code defining the host name and
contatning information about whether the code is run in the dev-server
or not.
This enables us to use different host names for development and
production - which will be necessary for the next commit where I add a
content switcher that compares the value of `window.location.host` with
the value definde in the `HOSTNAME` env var. Without a dynamically
defined host name openening the app in development mode will only show
you the redirect message.
Furhtermore, I can restrict the hot-reloader to only run in development
mode and not in production.