Added preview configuration
With this commit I added the option to preview the website with
production-like bundled code.
I added a run script in the `package.json` called 'preview'. This script
bundles the code just like the 'build' script, but it sets the now new
`PREVIEW` flag. This flag changes the host name in the production config
to the local firebase host name. After that it automatically starts
`firebase serve` and you can look at the app as it would run in
production.