Installed FireBase
In this commit I installed the FireBase package and added a `GameLoader`
component that connects to the FireStore. So far nothing is written or
read from the store. This will change in the next commits.
This serves as a base for persisting games, so that on reload of the
page your progress is not lost.