Persistent games
In this commit I created the first prototype for persisting games in our
FireStore backend.
I added the packages `js-cookie` and `uuid` for storing the game id in a
Cookie and for generating new game ids in a safe manner.

I merged `GameLoader.jsx` into `GameLoader.jsx` as I think they belong
together and won't really benefit from individual component files. Those
files would only clutter my src folder.
The `Game` component is now a mere functional component, nice. All the
state is handled by the `GameLoader`.

I extracted the database connection initialization into the `db.js`
file for it to be imported by every file that needs it.

In implementing the database synchronization I really mengled with the
current `GameState`. It is not used in a object orientated manner
anymore, but rather in a functional way. The overhead caused by creating
new GameState objects throughout the code really depreciates the
performance notably. I will try to fix this in future commits. I am
planning on moving to a function way of dealing with the game state
using immutable js as well.

The way games are loaded an synced:
The first thing that gets cared about is the game id. If there is no
game id found in the cookie, one is generated using the uuidv4 standard.
With this id we try loading the game in a transaction. If no game is
found a new one is created in the database, using the default start
values for the cells. (Here we have a workaround going on, that updated
the game, even if it already exists, because otherwise FireStore will
throw an error.) With the game initialized in the database, or just
already existing, we move on to registering the snapshot listener for
this particular game. The field is then rendered for the first time.
When the user now connects two numbers, the GameLoader updates the cells
representing the game state in the database. A new snapshot is sent and
our snapshot listener, which we registered in the beginning, updated the
`GameLoader` state with the new cells. The new cells are passed down to
the `Game` component and the changes get visible to the user.

Possible caviats:
With bad internet connection this will be probably unplayable.