Ability to start a new game after you won
In this commit I added a button to start a new game, when the current game is
won. This was added because otherwise you would just sit there with your 'You
Won!' message and could not continue playing.

I updated the `Game.jsx` quite a bit for that. I separated the game state
synchronizing logic from the display logic. The sync logic lives in the
`GameLoader` component, while displayment and game state management is handled
by the `Game` component in the same file.
I am much happier with the separation like this, although I feel it is not
quite perfect yet. The future will show if it can stand the test of time.

I also took the detour of making a fancy animation for the button to show up.
While doing so I also bumped up the timing and introduction of the 'You Won!'
message. It now waits for the field to finish its collapsing animation, then
appears, and then the 'Extend' button changes over to 'New Game'.