Added option to give up and start over
In this commit I botched the UI a bit by slamming in a "New Game" button, just
below the title. Looks not thaat nice. I would like to solve this in a different
way, but that would require a lot of changes which would be out of scope for
this branch.

So ugly - but functional - it is for now.