Button component
In this commit I extracted the buttons used throughout the ui into a separate
component. As I want to have one SASS file per component, I mainly extracted the
button into its own component, so that I can have a SASS file for it as well.
I didn't like how I was using mixins for the button styles and how I extended
the mixin for every button in the ui, even when there was no additional styling
needed.