Official Logo
Yes! We finally have an official logo!

In this commit I did everything to replace the old text-based name with a custom
made logo. This will hopefully improve the recall factor of my game and also
help in a technical way, as a text-based logo is prone to fonts not loading or
text being rendered incorrectly breaking the logo. Also I hope I can create a
nicer menu this way or any menu at all as at the moment of writing this the game
has no menu, just a 'New Game' button floating around with no real place to go.

I added the logo as vector graphic. Installed the `file-loader` package and
updated the `webpack.config.js` to be able to use the logo in my code. Doing so
I moved some packages to the `devDependencies` that belong there.
I threw out the text-based title, added the logo, tweaked the scss a bit.
The Sarpanch font was removed and the `RedirectMessage` adapted to use the
normal font instead. Doing so I also threw out the `StyledName` and `StyledLink`
components as we just write the name of the game now, without having a special
font applied to it.