Folable Menu
In this commit I implemented a small foldable menu, to provide a better place
for the 'New Game' button and thus have it not break the layout anymore.

This was quite the act. I moved the code for the header into the `Game.jsx` and
`RedirectNotice.jsx` component, as these components now have different headers,
one with the foldable menu, the other one without. (It would not make sense to
start a new game when you are just seeing the redirect notice.)
Moving the header with the foldable menu into the GameLoader component had the
advantage that I didn't need to install a centralized state management. I can
just directly call the function to start a new game inside the GameLoader when
the user clicks on the 'New Game' button in the foldable menu. This however
makes the GameLoader quite the extensive component as it more or less managaes
all the state there currently is. I hope to move to a centralized state
management like redux in the future.

To make for nicer transitions / loading information I build in a distinction
between initial load `appLoading` and a consecutive load `newGameLoading`. This
way I can easily implement a loading screen in the future. But when the user
starts a new game, not everything is hidden behind a loading screen again. Makes
for the smoother experience IMO.

I also put in some timeout logic to hide away the menu after the app has loaded
and after the user started a new game. This way the menu collapses on its own to
be in a more aesthetic state. The menu is visible some time after the load as
this tells the user where to find it. I didn't want to label anything or put
some hamburger menu anywhere as it would break the clean design of the game.

I removed the old 'New Game' button and its scss code.