Clicky buttons
In this commit I updated the transition on buttons when they get clicked. They
now not turn blue, but get scaled down a bit to appear as if they were pressed
in. This makes for a nicer feedback when the user clicks a button.

I also implemented a new style of button, an invisible one. This one can be used
to wrap elements without a buttony appearance. I used this to wrap the logo.
This makes the logo look neat, while it still has all the benefits of a button.
This should help browsers to notice this is a clickable element, this might be
especially useful for mobile devices.