Reworked Button change anmation when game is finished
In this commit I reworked the 3d button switch animation where the 'Extend'
button rotates over to become the 'New Game' button. This animation was very
crudely styled, so that some of the recently added button styles would get
overwritten.

I now extracted the code into an individual component which lives with the other
buttons in the `Btn.jsx`. I also separated the content from the animated
elements. This allows me to add a button into the animation without the button
being affected in any way. And this means the click or tap animation now works
on the contained buttons as well.

I also cleaned the styles a bit and improved some values here and there. Quite
pleased with the result.

I still get the bug, where the button does not get rendered in FireFox unless I
click or move the mouse near to it.