Loading Screen for initial loading of the game
In this commit I added a loading screen to hide away everything until the game
is fully loaded.
For that I created an inverted version of that logo to be used on a small cube
rotating randomly while things are loading.
I also created a new component that contains the loading indicator as well as a
loading screen and added css for both components.
I wired up those new components in the `Game` component, adding a small timeout
to have the screen fade out when loading is finished.