Added Anonymous Authentication and security rules
In this commit I added anonymous authentication to the game. This way I can
employ strict security rules for the documents in the FireStore.

I also added security rules for the FireStore. These will protect the data from
being tampered with as the access to the store is given with opening the JS
debugging console.

The rules are:
- you need to be authenticated in order to create a game
- only you can access the games you own.
- only you can update or write to the games you own, but only when you are
keeping the owner id the same. You cannot change the owner id.