Extended WebPack config for GitHub Pages and Conu.app
In this commit I improved the webpack configuration so that I can set a flag
whether to bundle the code for GitHub pages or for the official website. So far
this is only important to adapt the paths as we need different on the GitHub
pages an the official website.

I also reworked the npm scripts to allow for controlling which flavor to build.

Finally, I once ran the build script and updated the code for the GitHub pages.