No longer rely on Cookies for storing gameID
In this commit I moved from using cookies for storing the users current game id
to querying the most recent game the user created. This now relies on the
firebase anonymous authentication providing a constant user id. But as far as I
read that up on the internet this should be the case.

I more or less crudely stuffed everything in the `Game.jsx` component to have it
work. I want to extract that logic into a separate class in the next commit.