Moved URL switching into root component
I think I overdid it there with splitting components, so I moved the URL based
content switching logic back into the root component. I think this makes the
logic more accessible and thus easier to change the components to be rendered
according to the URL. Having to open just another file, just to see which
components get rendered when can be quite tedius.