Dummy redux implementation
In this commit I started adding redux to the game as a mean of state management.
I also added immutable-js for having an immutable state in my redux store.

So far only the user authentication is handled via redux and a dummy `AppLoader`
component is used to diplay the state change.

This is the groundwork for a generally better state management as promised when
I reworked the loading of the game state without relying on cookies.