Connecting game loading to redux
In this commit I connected the loading of the game to the redux state. While I
like that the game state now has a place to live, I don't feel comfortable with
putting things like the unsubscribe method and the gameRef in the redux store. I
feel those objects / functions are not really suitable to be put there. But I
thought better there than somewhere else. So everything is at one central
point. This is an experiment. Lets see how it goes.

I added `game.js` to care about the game state, just like `user.js` cares about
the users state. I added dispatchers and state mapping to the `AppLoader` and
added the game reducer to the root one.

The logic for the game initialization was taked from the `GameLoader` component
and only slightly modified to work with dispatching actions rather than setting
a components state.