Adding CSSTransition to the loading-screen
In this commit I wrapped the loading-screen in a `CSSTransition` component to
reduce the complexety of animating it fading out. This fading is now no longer
realized via the state of the surrounding component, but independently managed
by the loading-screen itself.
This makes using the loading-screen much easier and your component less
cluttered.