Extracted Header into own component
In this commit I extracted the header logic from the `GameLoader` into its own
component. This reduces the complexity of the GameLoader even more, slowly
dissassembling it until it is no longer needed as a state manager for the `Game`
component.
I employed the same animation technique as with the loading screen, this way I
didn't need any state handling in the header component.
I also found a neat way of setting and clearing the timeout for delayed collapse
of the menu. The timeout is manages inside the redux header
reducer/action-creator `redux/header.js` as well as returned to the component on
call of `collapseMenuWithTimeout` allowing the component to individually clear
the time out as well. (For example in case of an unmount.)
I decided against using the header for the `RedirectNotice` as well, as I
thought this would only complicate the component with no real benefit.