Functional Game State
In this commit I reworked the `GameState` to be fully functional and no longer
object orientated. This will make the it much more usable in our new
redux state management, as we manage our state there and just need to verify the
validity of our state or of the moves te player makes. This ist why the
`GameState` was renamed to `GameLogic`.