Removing the user object from the reduc state
In this commit I followed up on cleaning out the redux state, removing the
firebase user object from the state. I think this it is not neccessary to have
it there, maybe even dangerous as it can lead to errors. (Just like the game
reference no longer working, when put in the redux state.)
I think I'll try to keep only the things in the state that are directly needed
for the app to render.