Removed the document snapshot listener
In this commit I removed the document snapshot listener. The more I work with
redux the more I feel this is not the way to go for updating the games state. I
now plan on having clearly defined actions that will asynchronously update the
state in the database, while having multiple actions for the redux store
updating the redux state according to the progress. They will work the same as
the game initialization, where there are three different actions indicating the
game initialization progress.
This will also allow me to implement a small loading animation on the button (or just
disabling the button) while the new state is persisted in the firestore. Sure,
this would also work with the snapshot-listener approach, but I would not find
that as elegant and it might be more prone to error when thinking about multiple
user playing on one game. (Not sure if that will ever be a thing though.)

Additionally, I cleaned up my nested promise mess. Now things are neat and tidy.