Restored cell selecting and pairing functionality
In this commit I extended the redux game reducer to be able to select and pair
cells again.

I added the dispatch to the `Cell` component, so that it can directly trigger
the state change, when the user clicks on it.

I removed the cell selecting / pairing logic from the `Game` component, further
reducing its complexity and removing ecerything it does no longer need to care
about.

Finally, I extended the game reducer to accomodate for two new actions, one when
a cell is selected, and one when two cells got paired and the changes were
persisted in the database. There are more actions, one for when the pairing of cells
is started and one for when the pairing fails due to a database problem or invalid
indices. These are not yet used in the state however. I think about using those
to further improve the animation of cells getting selected and then paired.
Besides that, the logic is basically the way it used to be in the `Game`
component. Just the updating is done without the snapshot listener, reacting to
the resolving of the promise of the firestores update function instead. In my
tests this seems to impact performance quite a lot, making the game feel a log
less snappy. I'll see if I can address this in the future.