Repaired extend field functionality
In this commit I extended the redux game reducer to handle the extension of the
playing field. Just as with the commit before where I repaired pairing and
cell selection logic, I used three actions that indicate starte, success and
failure of the extension of the game field. With success also implying the
updated game state being persisted in the firestore.

Then I wired things up in the `Game` component and removed the old extension
function from there.