Highlight pairing cells
In this commit I tried to make the cell pairing feel a bit smoother by adding a
new `pairing` state in which the Cell is theoretically paired, but we are still
waiting for the changes to be persisted in the firestore.

I added a new set in the redux game state for this. This set contains indices of
paired cells that are currently waiting for their changed status to be updated
in the firestore.
On the `Cell` components side we pass in a new flag prop `pairing` derived from
whether the cells index is contained in the aforementioned set. When true, the
cell applies a new css class givin git a green background.

I also stopped resetting the selected cell, whenever two paired cells were
persisted in the database, as by this time the user might have already moved on
and selected their next cell to pair up.