Make game id cookie persistent
In this commit I added an expiry time to the game id cookie. Otherwise it will
be a session cookie and expire as soon as the user closes its browser. (Never
tested that as my browser is always running when I develop, yeah I know.)
With the expiry set to one year, it should be fine. I thought about taking an
even longer interval, but in one year you probably cleared your cache / history,
uninstalled your browser, upgraded to a new device or what not, and probably
have forgotten about this game. So one year should be plenty for the ones who
stick around and play it frequently.