Game reducer can start new game
In this commit I reimplemented the ability to start a new game in the game
reducer. With this in place, the game can do all the things it could do before I
started moving to redux again.

I will now focus my attention on cleaning up the code and tweak a few things, so
that the game behaves just like before again.