Separated loading animations for init and new game
In this commit I reintroduced the two loading animations that were used
previously. One for when the whole app is loading for when the site is loaded
initially and the other for when the user starts a new game.
The former is appearing on the whole screen functioning as a loading screen,
while the latter is more of a loading animation which appears where the game
field is when a new game is loading. The difference thus is, that you can still
see the header with the latter.
I find this distinction important as I think it makes for a nicer user
experience. Havin the whole app vanish behind a loading screen, just because you
started a new game almost feels like you broke something.

For that I introduced a new flad in the reudx state, that indicates if the app /
game is initially loading, or not. It initially is set to true and on the first
successfull loading of the game it switches to false and then never comes back
on again.

While doing so, I also disabled the "New Game" button, when the game is
currently loading.