Removed `Field` structure from game logic
Before we had this 'Field' structure in the game logic that would represent the
game state ( the cells ) in an 2 dimensional array. This way you could access
rows and cells of those rows and get information about them.
This structure also represented the way, the field was rendered. Row by row and
then cell by cell. We need to render the rows individually, as only then we can
collapse them with an animation. (Rendering only the cells will cause the
collapse animation to not work properly, shifting the following cells in the
progress. )

This structure is not necessarily bad, but with moving to redux state
management, it is no longer needed. Because now each row and cell directly gets
all their information from the redux state.
With that this structure is just a bulky overhead that isn't needed anymore.
But more importantly, it is blocking the updating of the field. When a row is
emptied of pairable cells it won't get removed, cluttering the game field.

With this commit I completely removed the logic for generating this structure
from the `GameLogic` and replaced it with functions that return indices for rows
and cells of rows in a functional way.

I would assume this also brings performance benefits, but I currently have no
means of testing that.

I updated the `Field` and `Row` component to directly use indices instead of the
aforementioned 'field' structure.