Refined header menu animation
In this commit I reimplemented the logic for collapsing the menu after the user
started a new game. As it was one of the last pieces missing afther the redux
state implementation.

While doing so, I also refinde the way the menu behaves. Basically, It reacts to
the users input, staying open until a new game is loaded, even when it was
scheduled to collapse before. Or stopping the closing animation, if the user
clicks on the menu, just when it was about to collapse.
For this a new `showMenu` action was created. The state of the menu was renamed
to `menuCollapsed` to avaoid conflicts.