Fixed last row being displayed along with won-message
In the logic for deiding for whether the last row should be displayed or not, I
forgot to add a check to see if the game was finished. This way the last row
could show up, when not completely filled along with the won-message.

I added a check that prevents the last row being displayed when the game is
finished or in the words of the check: when there are no other active rows.

While doing so, I also realized, that the Field was still being displayed when
the page was reloaded and a finished game was loaded. I added a `onMountEnter`
flag to the surrounding CSSTransistion component to prevent the Field from being
rendered, when the in prop is set to false.