No won-message animation on page load, fix new game button
When loading the page and your current game is already won, the won-message will
still be animated. I think this looks a bit strange and prefer a static won
message. It was already animated when the player won the game, so it just makes
sense to me that you find it just like this when you reload the page.

This allowed me to use the CSSTransitions classes and thus I no longer need to
use an animation in the CSS. Much nicer IMO.

Furthermore, I set the won-message entered flag depending on whether the game is
finished to have the new-game button being displayed, when the app is loaded
with a won game.