Deploy script leaves `/docs` unchanged
In this commit I tweaked the deploy script so that it only builds the
`official-website` flavour of the app.
This way, running the deploy script will no longer build the `github-pages`
falvour as well, leaving the files in `docs` untouched. So after a deploy you no
longer need to `git checkout` the changed bundles in the `docs` folder.