Added deploy job and workflow
In this commit I added config for a deploy job and a corresponding workflow.
This will serve as a base for automatic deployment on releases on the master
branch. The workflow already has filters in place, which should just do that,
but I couldn't test them yet.

However, I am not yet happy with the configuration. I think I want to have the
jobs interconnected by artifacts, making them faster and less redundand.