Implemented release workflow
In this commit I implemented a release workflow in the CircleCI config. This
workflow replaces the deploy workflow which was my first attempt of automated
releases.

The workflow builds, tests and then deploys the code for all commits on the
master with a tag starting in `v.`.
I am using workspaces to persiste the built code and config from the `build` job
for the `deploy` job.
I am using plain firebase cli commands for the deploy. This allows me to remove
the deploy script from my `package.json`, which is a good thing as this way no
one is tempted to deploy using the npm script.