Updated FireStore indices
In this commit I updated the FireStore indices configuration to the as same as
the indices that are live at the moment.
These are needed for the new way of fetching games which no longer relies on the
game id being present in a cookie but on selecting the most recent game for the
currently logged in user. For such queries we need the indices.
In order to not break anything on deploy, I updated the indices configuration
with this commit.