Make deploy depend on build and test
I made a mistake when configuring the release workflow. I made the deploy only
dependent on the build job. This way the deploy job was dierectly run after the
build without any testing what so ever.

I now fixed that to make sure we only deploy tested code - well as if we have
this much test anyway.