Improved CircleCI config
With the last release, I noticed the deploy job actually running before the tag
was set. So I dug into the CircleCI documentation again and think I now have it
right.

I want to have the `test` workflow to run tests on all branches except for the
master branch. And then I have my `deploy_on_release` workflow that should only
run when there is a release on the master branch.

I don't like the idea of having parts of your workflow being executed depending
on the branch or tag. So I applied the same filters to all jobs in the each
workflow. Hopefully, this will keep my CircleCI job history clean and easy to
read / unerstand.