Really releasing only on tags
I think I now have got it. These filters are not working in an 'and' fashion,
but rather in an 'or' fashion. So by ignoring all branches - as the
documentation said, but I wouldn't listen - I now should have releases that only
trigger when I add a tag to a commit.