Converted Field component to hooks
With this commit the Field component is converted from using HOCs to using hooks
for using the redux state. The conversion was relatively straight forward. Only
thing that really chagned is that we are now using an immutable.js List for
storing the active rows. (The rows that contain pariable numbers and thus need to
be displayed.) This way we can simply provide the immutable.js `is` function as the
second argument to the `useSelector` function for comparing the states and
determining whether a rerender is needed. This prevents rerenders, when cells
are paried without any rows changing. That is why the `getActiveRowIndices`
function now returns a list, rather than an array.

Furthermore, I adapted the naming of the `activeRows` variable, to refelct the
functions name and thus be more coherent.